requirements.txt
Django==2.1.2
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.5
requests-html==0.9.0
schedule==0.5.0
